move order code from async.py to order.py
