+++ dask/core.py
@@ -216,6 +216,21 @@ def get_dependencies(dsk, task, as_list=False):
     return rv if as_list else set(rv)
 
 
+def get_deps(dsk):
+    """ Get dependencies and dependents from dask dask graph
+
+    >>> dsk = {'a': 1, 'b': (inc, 'a'), 'c': (inc, 'b')}
+    >>> dependencies, dependents = get_deps(dsk)
+    >>> dependencies
+    {'a': set([]), 'c': set(['b']), 'b': set(['a'])}
+    >>> dependents
+    {'a': set(['b']), 'c': set([]), 'b': set(['c'])}
+    """
+    dependencies = dict((k, get_dependencies(dsk, k)) for k in dsk)
+    dependents = reverse_dict(dependencies)
+    return dependencies, dependents
+
+
 def flatten(seq):
     """
